centralized life cycle logic
