+++ ingestion/src/metadata/ingestion/source/database/bigquery/metadata.py
@@ -34,7 +34,6 @@ from metadata.generated.schema.entity.data.databaseSchema import DatabaseSchema
 from metadata.generated.schema.entity.data.storedProcedure import StoredProcedureCode
 from metadata.generated.schema.entity.data.table import (
     IntervalType,
-    Table,
     TablePartition,
     TableType,
 )
@@ -54,7 +53,6 @@ from metadata.generated.schema.type.basic import EntityName, SourceUrl
 from metadata.generated.schema.type.tagLabel import TagLabel
 from metadata.ingestion.api.models import Either
 from metadata.ingestion.api.steps import InvalidSourceException
-from metadata.ingestion.models.life_cycle import OMetaLifeCycleData
 from metadata.ingestion.models.ometa_classification import OMetaTagAndClassification
 from metadata.ingestion.ometa.ometa_api import OpenMetadata
 from metadata.ingestion.source.connections import get_test_connection_fn
@@ -223,6 +221,7 @@ class BigquerySource(
         # list of all project IDs. Subsequently, after the invokation,
         # we proceed to test the connections for each of these project IDs
         self.project_ids = self.set_project_id()
+        self.life_cycle_query = BIGQUERY_LIFE_CYCLE_QUERY
         self.test_connection = self._test_connection
         self.test_connection()
 
@@ -541,38 +540,6 @@ class BigquerySource(
             for temp_file_path in self.temp_credentials_file_path:
                 if os.path.exists(temp_file_path):
                     os.remove(temp_file_path)
-
-    def yield_life_cycle_data(self, _) -> Iterable[Either[OMetaLifeCycleData]]:
-        """
-        Get the life cycle data of the table
-        """
-        try:
-            table_fqn = fqn.build(
-                self.metadata,
-                entity_type=Table,
-                service_name=self.context.database_service,
-                database_name=self.context.database,
-                schema_name=self.context.database_schema,
-                table_name=self.context.table,
-                skip_es_search=True,
-            )
-            table = self.metadata.get_by_name(entity=Table, fqn=table_fqn)
-            if table:
-                yield from self.get_life_cycle_data(
-                    entity=table,
-                    query=BIGQUERY_LIFE_CYCLE_QUERY.format(
-                        database_name=table.database.name,
-                        schema_name=table.databaseSchema.name,
-                    ),
-                )
-        except Exception as exc:
-            yield Either(
-                left=StackTraceError(
-                    name="lifeCycle",
-                    error=f"Error Processing life cycle data: {exc}",
-                    stackTrace=traceback.format_exc(),
-                )
-            )
 
     def _get_source_url(
         self,
